Put the LedVu component in its own file.
