+++ CMakeLists.txt
@@ -72,6 +72,7 @@ clap_juce_extensions_plugin(TARGET elastika-filter
 
 
 target_sources(elastika-filter PRIVATE
+  src/led_vu.cc
   src/sapphire_lnf.cc
   src/ElastikaProcessor.cpp
   src/ElastikaEditor.cpp
