+++ src/led_vu.cc
@@ -1,0 +1,12 @@
+#include "led_vu.h"
+
+namespace sapphire
+{
+
+void LedVu::paint(juce::Graphics &g)
+{
+    float diam = std::min<float>(getWidth(), getHeight()) - outerLine;
+    juce::LookAndFeel_V2::drawGlassSphere(g, 0, 0, diam, col, outerLine);
+}
+
+} // namespace sapphire
